+++ packages/nimble-components/src/select/tests/select-opened-matrix.stories.ts
@@ -2,7 +2,10 @@ import type { Story, Meta } from '@storybook/html';
 import { withXD } from 'storybook-addon-xd-designs';
 import { html, ViewTemplate } from '@microsoft/fast-element';
 import { createRenderer } from '../../utilities/tests/storybook';
-import { createMatrix, themeWrapper } from '../../utilities/tests/matrix';
+import {
+    backgroundStates,
+    singleThemeWrapper
+} from '../../utilities/tests/matrix';
 import '..';
 
 const metadata: Meta = {
@@ -23,7 +26,7 @@ export default metadata;
 const positionStates = [
     ['below', 'margin-bottom: 120px;'],
     ['above', 'margin-top: 120px;']
-];
+] as const;
 type PositionState = typeof positionStates[number];
 
 // prettier-ignore
@@ -39,6 +42,39 @@ const component = ([
     </nimble-select>
 `;
 
-export const selectOpenedThemeMatrix: Story = createRenderer(
-    themeWrapper(createMatrix(component, [positionStates]))
+const [
+    lightThemeWhiteBackground,
+    colorThemeDarkGreenBackground,
+    darkThemeBlackBackground,
+    ...remaining
+] = backgroundStates;
+
+if (remaining.length > 0) {
+    throw new Error('New backgrounds need to be supported');
+}
+
+export const selectBelowOpenLightThemeWhiteBackground: Story = createRenderer(
+    singleThemeWrapper(component(positionStates[0]), lightThemeWhiteBackground)
+);
+
+export const selectAboveOpenLightThemeWhiteBackground: Story = createRenderer(
+    singleThemeWrapper(component(positionStates[1]), lightThemeWhiteBackground)
+);
+
+// prettier-ignore
+export const selectBelowOpenColorThemeDarkGreenBackground: Story = createRenderer(
+    singleThemeWrapper(component(positionStates[0]), colorThemeDarkGreenBackground)
+);
+
+// prettier-ignore
+export const selectAboveOpenColorThemeDarkGreenBackground: Story = createRenderer(
+    singleThemeWrapper(component(positionStates[1]), colorThemeDarkGreenBackground)
+);
+
+export const selectBelowOpenDarkThemeBlackBackground: Story = createRenderer(
+    singleThemeWrapper(component(positionStates[0]), darkThemeBlackBackground)
+);
+
+export const selectAboveOpenDarkThemeBlackBackground: Story = createRenderer(
+    singleThemeWrapper(component(positionStates[1]), darkThemeBlackBackground)
 );