+++ angular-workspace/projects/ni/nimble-angular/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Tue, 22 Mar 2022 14:07:45 GMT and should not be manually modified.
+This log was last generated on Tue, 22 Mar 2022 20:59:12 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 4.2.1
+
+Tue, 22 Mar 2022 20:59:12 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v7.2.3
 
 ## 4.2.0